Start refactoring scopes
